Promote InstanceId to package chisel3
Deprecate DataMirror alias in chisel3.experimental.

Move package object internal from RawModule.scala to
internal/package.scala.
