+++ core/src/main/scala/chisel3/Annotation.scala
@@ -3,8 +3,8 @@
 package chisel3.experimental
 
 import scala.language.existentials
-import chisel3.internal.{Builder, InstanceId, LegacyModule}
-import chisel3.{CompileOptions, Data, RawModule}
+import chisel3.internal.{Builder, LegacyModule}
+import chisel3.{CompileOptions, Data, InstanceId, RawModule}
 import firrtl.Transform
 import firrtl.annotations._
 import firrtl.options.Unserializable
